requirements.txt
pygame
pyserial
plotly